+++ crates/minecraft_packets/src/play/data/light_data.rs
@@ -1,0 +1,18 @@
+use minecraft_protocol::prelude::*;
+
+#[derive(PacketOut, Default)]
+pub struct LightData {
+    sky_light_mask: BitSet,
+    block_light_mask: BitSet,
+    empty_sky_light_mask: BitSet,
+    empty_block_light_mask: BitSet,
+    sky_light_arrays: LengthPaddedVec<Light>,
+    block_light_arrays: LengthPaddedVec<Light>,
+}
+
+#[derive(PacketOut, Default)]
+pub struct Light {
+    /// Length of the following array is always 2048
+    /// There is 1 array for each bit set to true in the light mask, starting with the lowest value. Half a byte per light value. Indexed ((y<<8) | (z<<4) | x) / 2 If there's a remainder, masked 0xF0 else 0x0F.
+    block_light_array: LengthPaddedVec<i8>,
+}