+++ crates/minecraft_packets/src/play/data/mod.rs
@@ -1,2 +1,5 @@
+pub mod chunk_data;
 pub mod chunk_section;
+mod encode_as_bytes;
+pub mod light_data;
 pub mod palette_container;