+++ crates/minecraft_protocol/src/data_types/bit_set.rs
@@ -1,18 +1,17 @@
 use crate::prelude::{EncodePacket, LengthPaddedVec};
+use macros::PacketOut;
 use pico_binutils::prelude::{BinaryWriter, BinaryWriterError};
 use protocol_version::protocol_version::ProtocolVersion;
 
-#[derive(Debug, Default)]
+#[derive(Default, Clone, PacketOut)]
 pub struct BitSet {
     data: LengthPaddedVec<i64>,
 }
 
-impl EncodePacket for BitSet {
-    fn encode(
-        &self,
-        writer: &mut BinaryWriter,
-        _protocol_version: ProtocolVersion,
-    ) -> Result<(), BinaryWriterError> {
-        writer.write(&self.data)
+impl BitSet {
+    pub fn new(data: Vec<i64>) -> Self {
+        Self {
+            data: LengthPaddedVec::new(data),
+        }
     }
 }