+++ pico_limbo/src/handlers/configuration.rs
@@ -138,7 +138,7 @@ pub fn send_play_packets(
     let packet = build_login_packet(protocol_version, server_state.spawn_dimension())?
         .set_game_mode(game_mode.value())
         .set_view_distance(view_distance)
-        .set_hardcore(server_state.is_hardcore());
+        .set_hardcore(protocol_version, server_state.is_hardcore());
     client_state.queue_packet(PacketRegistry::Login(Box::new(packet)));
 
     let (x, y, z) = server_state.spawn_position();
@@ -153,10 +153,12 @@ pub fn send_play_packets(
         client_state.queue_packet(PacketRegistry::Commands(packet));
     }
 
-    if protocol_version.is_after_inclusive(ProtocolVersion::V1_20_3) {
-        // Send Game Event
-        let packet = GameEventPacket::start_waiting_for_chunks(0.0);
-        client_state.queue_packet(PacketRegistry::GameEvent(packet));
+    if protocol_version.is_after_inclusive(ProtocolVersion::V1_19) {
+        if protocol_version.is_after_inclusive(ProtocolVersion::V1_20_3) {
+            // Send Game Event
+            let packet = GameEventPacket::start_waiting_for_chunks(0.0);
+            client_state.queue_packet(PacketRegistry::GameEvent(packet));
+        }
 
         // Send Chunk Data and Update Light
         let biome_id = get_void_biome_index(protocol_version).ok_or_else(|| {
@@ -171,9 +173,8 @@ pub fn send_play_packets(
         let packet = SetCenterChunkPacket::new(center_chunk.0, center_chunk.1);
         client_state.queue_packet(PacketRegistry::SetCenterChunk(packet));
         for (chunk_x, chunk_z) in chunk_positions {
-            let packet =
-                ChunkDataAndUpdateLightPacket::new(protocol_version, chunk_x, chunk_z, biome_id);
-            client_state.queue_packet(PacketRegistry::ChunkDataAndUpdateLight(packet));
+            let packet = ChunkDataAndUpdateLightPacket::void(chunk_x, chunk_z, biome_id);
+            client_state.queue_packet(PacketRegistry::ChunkDataAndUpdateLight(Box::new(packet)));
         }
     }
 
@@ -299,6 +300,14 @@ mod tests {
         assert!(matches!(
             client_state.next_packet(),
             PacketRegistry::Commands(_)
+        ));
+        assert!(matches!(
+            client_state.next_packet(),
+            PacketRegistry::SetCenterChunk(_)
+        ));
+        assert!(matches!(
+            client_state.next_packet(),
+            PacketRegistry::ChunkDataAndUpdateLight(_)
         ));
         assert!(matches!(
             client_state.next_packet(),